presence-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
kafka-python==2.0.2
redis==5.0.4
pydantic==2.7.1
python-jose==3.3.0
python-multipart==0.0.9